Version Upgrade
# Description
Upgraded to Godot 4.4.

Project was reset to include UIDs. Currently not being used for preloading.

Slightly reduced the player hitbox according to feedback.

Modified the shooter to be able to customise the bullet size.

Adjusted the level three due to being able to stand in the middle and avoid being hit with the new hitbox.

Removed outline when interacting with the level transition buttons.

Renamed "death_trigger" to "DeathTrigger" to match the naming convention of the other script files.