Refactoring
# Description
Reorganised scripts into categories

Removed the button scripts that would provide an action for clicking on them for each level loading action (start game and skip buttons). It was replaced with a singular node script that contains all the actions that the button signal can select based on the most relevant action.

Renamed "PlayerSave" to "SaveFile" for more accurate depiction.

Added GameSettings autoload script for the game settings that will be added.