Keycodes & UI
# Description
Settings button currently is being used for testing. When clicked it will start listening for input and display the key and associated texture after a valid keycode is pressed. Not all textures are provided, defaulting to cursor in some.

Updated the `InputActionData` to be able to retrieve from the project settings the keycodes associated with the given action. Additionally it can now verify that the selected name is registered in the `InputMap` inside the project settings.