Settings Key Rebind
# Description
Settings UI is now able to listen for only 1 key at a time and the user is able to select which input they want to listen for by clicking on the button next to their desired input to rebind. Once a valid key is pressed, the input is either changed or ignored if already bind to another input. Player must confirm their selection by clicking on the confirm changes button or cancel them by clicking on the close button.